Fix some lint warnings and suppress the others (#8395)
### Description

Fix some lint warnings and suppress the others. This makes builds easier
because IntelliJ no longer jumps to those warnings (and it prevents more
from showing up).

Initial prototype: Claude, $1.99

### Checklist
<!-- 
To help us keep the issue tracker clean and work as efficient as
possible,
  please make sure that you have done all of the following.
You can tick the boxes below by placing an x inside the brackets like
this: [x]
-->
- [x] I have read the contribution guidelines:
https://github.com/AntennaPod/AntennaPod/blob/develop/CONTRIBUTING.md#submit-a-pull-request
- [x] I have performed a self-review of my code, going through my
changes line by line and carefully considering why this line change is
necessary
- [x] I have run the automated code checks using `./gradlew checkstyle
lint spotbugsPlayDebug spotbugsDebug`
- [x] My code follows the style guidelines of the AntennaPod project:
https://antennapod.org/contribute/develop/app/code-style
- [x] I have mentioned the corresponding issue and the relevant keyword
(e.g., "Closes: #xy") in the description (see
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue)
- [x] If it is a core feature, I have added automated tests

+++ ui/echo/src/main/java/de/danoeh/antennapod/ui/echo/screen/QueueScreen.java
@@ -87,7 +87,7 @@ public class QueueScreen extends EchoScreen {
                         }
                         if (item.getMedia() != null) {
                             long itemTimeLeft = item.getMedia().getDuration() - item.getMedia().getPosition();
-                            queueSecondsLeft += itemTimeLeft / playbackSpeed;
+                            queueSecondsLeft += (long) (itemTimeLeft / playbackSpeed);
                         }
                     }
                     queueSecondsLeft /= 1000;
